Fixes a few ships' TEG setups (#3075)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Removes the disposals chutes some TEG setups have for you to throw lit
welding tools into, and replaces them with proper igniters.

## Why It's Good For The Game

Having a disposals chute to throw lit welding tools into was a stop-gap
solution to igniters not working, which was fixed over a year ago in
#1675.

## Changelog

:cl:
add: Adds an igniter and button to the TEG of every ship that was
missing one
del: Removed the welding tool TEG chute some ships had
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->