Ports 'Command bar typing indicators (client side html version)' (#3080)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
Directly ports https://github.com/tgstation/tgstation/pull/83081,
allowing typing indicators to trigger off of the command bar.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Typing indicators are there for a reason! It's helpful to know if
another player is cooking a message for thirty seconds or just staring
at you. However, a good chunk of our players exclusively use the command
bar to send messages, leaving us in the dark! This helps level the
playing field.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: typing indicators now trigger off of the command bar
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->